refactor: 강의 목록 렌더링 로직을 CourseListView 컴포넌트로 분리
- Home, Courses, Search 페이지의 중복된 강의 카드, 그리드, 페이지네이션 렌더링 로직을 CourseListView로 분리
- 각 페이지에서는 필터 로직 및 상태 관리를 유지하여 의도와 책임을 명확히 함
