+++ src/pages/Search.jsx
@@ -1,14 +1,11 @@
 import { useSearchParams } from 'react-router-dom';
 import { useState, useMemo } from 'react';
-import { Container, Grid, Center, Pagination, Title, Text } from '@mantine/core';
 import CategoryTabs from '../components/CategoryTabs';
 import FilterBar from '../components/FilterBar';
-import CourseCard from '../components/CourseCard';
-import CATEGORIES from '../data/courses.json';
+import CourseListView from '../components/CourseListView';
+import courses from '../data/courses.json';
 
-const ITEMS_PER_PAGE = 40;
-
-function Search() {
+const Search = () => {
     const [searchParams] = useSearchParams();
     const keyword = searchParams.get('s')?.toLowerCase() || '';
 
@@ -21,7 +18,7 @@ function Search() {
     };
 
     const filteredCourses = useMemo(() => {
-        return CATEGORIES.filter((course) => {
+        return courses.filter((course) => {
             const matchKeyword = keyword === '' || course.title.toLowerCase().includes(keyword);
             const matchDifficulty =
                 filters.difficulty.length === 0 || filters.difficulty.includes(course.level);
@@ -32,42 +29,19 @@ function Search() {
         });
     }, [keyword, filters]);
 
-    const totalPages = Math.ceil(filteredCourses.length / ITEMS_PER_PAGE);
-    const startIndex = (activePage - 1) * ITEMS_PER_PAGE;
-    const endIndex = startIndex + ITEMS_PER_PAGE;
-    const currentCourses = filteredCourses.slice(startIndex, endIndex);
-
     return (
         <>
             <CategoryTabs />
-            <Container size='xl' py='md'>
-                <Title order={2} mb='xs'>‘{keyword}’ 검색 결과</Title>
-                <Text mb='md' size='sm' c='dimmed'>
-                    {filteredCourses.length}개의 강의가 검색되었습니다.
-                </Text>
-
-                <FilterBar onFilterChange={handleFilterChange} />
-
-                <Grid gutter='lg'>
-                    {currentCourses.map((course) => (
-                        <Grid.Col key={course.courseCode} span={2.4}>
-                            <CourseCard {...course} />
-                        </Grid.Col>
-                    ))}
-                </Grid>
-
-                <Center mt='xl'>
-                    <Pagination
-                        value={activePage}
-                        onChange={setActivePage}
-                        total={totalPages}
-                        color='#00c471'
-                        withEdges
-                    />
-                </Center>
-            </Container>
+            <CourseListView
+                title={`‘${keyword}’ 검색 결과`}
+                description={`${filteredCourses.length}개의 강의가 검색되었습니다.`}
+                courses={filteredCourses}
+                activePage={activePage}
+                setActivePage={setActivePage}
+            />
+            <FilterBar onFilterChange={handleFilterChange} />
         </>
     );
-}
+};
 
 export default Search;
